Move validation rules to validation.go
